few header related changes.
